[ZW-5191] Title: Panic causes SQS connection to close on Dropped Tx
JIRA Link: https://kzencorp.atlassian.net/browse/ZW-5191
Reviewer: Oded
Comments: When a dropped transaction is found, the new_statushash is
Null. We use this as the groupId for sending the tx to SQS, the
groupId is needed otherwise the operation fails
(see: Dev Docs in rusoto_sqs/src/generated.rs:SendMessageRequest).

If a dropped transaction is found we now use a uuid:v4 string as the
groupId to prevent a panic